requirements.txt
Flask==0.12.2
Flask-Login==0.4.1
Jinja2==2.10
PyYAML==3.12
uWSGI==2.0.17
